# docs/swagger.yaml
openapi: 3.0.3
info:
  title: RtBrick BNG Blaster Conroller API
  description: >-
    RtBrick BNG Blaster Conroller API is the management API for multiple BNG Blaster instances.

    It allows to manage instances in the sense that new instances can be created, started, stopped 
    and also controlled via command socket of the BNG Blaster.
  contact:
    email: bngblaster@rtbrick.com
    url: https://github.com/rtbrick/bngblaster-controller
  version: 0.1.3
servers:
  - url: 'http://localhost:8001'
paths:
...
  /api/v1/schema:
    get:
      summary: BNG Blaster configuration JSON schema.
      description: >-
        Get the JSON schema used to render and validate the "New Instance" config editor in the web UI.
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                type: object
        404:
          description: not found, schema not available
          content:
            text/plain:
              schema:
                type: string
...
  /api/v1/instances/{instance_name}:
    get:
      summary: Status information of an instance.
      description: >-
        Get status information of an instance.
      parameters:
        - name: instance_name
          description: instance name of the bngblaster
          in: path
          required: true
          example: sample
          schema:
            type: string
      responses:
        200:
          description: ok, the status object
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum:
                      - started
                      - stopped
        204:
          description: no content, the instance was updated
        400:
          description: bad request, body not parsable
          content:
            text/plain:
              schema:
                type: string
        404:
          description: not found, if instance does not exist
          content:
            text/plain:
              schema:
                type: string
        412:
          description: precondition failed, if instance is running
          content:
            text/plain:
              schema:
                type: string
    put:
      summary: Create or update an instance
      description: >-
        The bngblaster instance will be created if there is not already one.
        If the instance already exists the configuration is replaced.
        The instance will not be started only prepared.
      parameters:
        - name: instance_name
          description: instance name of the bngblaster
          in: path
          required: true
          example: sample
          schema:
            type: string
      requestBody:
        description: The config file for the bngblaster instance.
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                oneOf:
                  - type: string
                  - type: integer
                  - type: object
            example:
              {
                "interfaces": {
                  "network": {
                    "interface": "bblB",
                    "address": "10.0.0.10",
                    "gateway": "10.0.0.2",
                    "address-ipv6": "fc66:1337:7331:8::10",
                    "gateway-ipv6": "fc66:1337:7331:8::1"
                  },
                  "access": {
                    "interface": "bblA",
                    "outer-vlan-min": 1,
                    "outer-vlan-max": 4049,
                    "inner-vlan-min": 1,
                    "inner-vlan-max": 4049
                  }
                }
              }
      responses:
        201:
          description: created, the instance was created
        204:
          description: no content, the instance was updated
        400:
          description: bad request, body not parsable
          content:
            text/plain:
              schema:
                type: string
        412:
          description: precondition failed, if instance is running
          content:
            text/plain:
              schema:
                type: string
        500:
          description: internal server error
          content:
            text/plain:
              schema:
                type: string
    delete:
      summary: Delete an instance.
      description: >-
        Deletes an instance. This does not stop a running instance, only cleans up everything if the instance is already stopped.
      parameters:
        - name: instance_name
          description: instance name of the bngblaster
          in: path
          required: true
          example: sample
          schema:
            type: string
      responses:
        204:
          description: no content, the instance was deleted
        412:
          description: precondition failed, if instance is running
          content:
            text/plain:
              schema:
                type: string
        500:
          description: internal server error
          content:
            text/plain:
              schema:
                type: string
  /api/v1/instances/{instance_name}/_streams:
    get:
      summary: List traffic streams of a running instance.
      description: >-
        Get a paginated, optionally filtered view of the streams reported by the
        bngblaster "stream-summary" control socket command. Used by the web UI's
        virtual-scrolling stream table, which only ever requests the slice of rows
        currently in (or near) its viewport.
      parameters:
        - name: instance_name
          description: instance name of the bngblaster
          in: path
          required: true
          example: sample
          schema:
            type: string
        - name: offset
          description: number of streams to skip
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          description: maximum number of streams to return
          in: query
          required: false
          schema:
            type: integer
            default: 50
            maximum: 500
        - name: session-id
          in: query
          required: false
          schema:
            type: integer
        - name: session-group-id
          in: query
          required: false
          schema:
            type: integer
        - name: flow-id
          in: query
          required: false
          schema:
            type: integer
        - name: flow-id-min
          description: lower bound of an explicit flow-id range; requires flow-id-max
          in: query
          required: false
          schema:
            type: integer
        - name: flow-id-max
          description: upper bound of an explicit flow-id range; requires flow-id-min
          in: query
          required: false
          schema:
            type: integer
        - name: name
          in: query
          required: false
          schema:
            type: string
        - name: interface
          in: query
          required: false
          schema:
            type: string
        - name: direction
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum:
              - verified
              - bidirectional-verified
              - pending
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/streamsResponse'
        404:
          description: not found, instance does not exist
          content:
            text/plain:
              schema:
                type: string
        412:
          description: precondition failed, if instance is not running
          content:
            text/plain:
              schema:
                type: string
        500:
          description: internal server error
          content:
            text/plain:
              schema:
                type: string
  /api/v1/instances/{instance_name}/_sessions:
    get:
      summary: List sessions of a running instance.
      description: >-
        Get a paginated, optionally filtered view of the sessions reported by the
        bngblaster "session-summary" control socket command. Used by the web UI's
        virtual-scrolling session table, which only ever requests the slice of rows
        currently in (or near) its viewport.
      parameters:
        - name: instance_name
          description: instance name of the bngblaster
          in: path
          required: true
          example: sample
          schema:
            type: string
        - name: offset
          description: number of sessions to skip
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          description: maximum number of sessions to return
          in: query
          required: false
          schema:
            type: integer
            default: 50
            maximum: 500
        - name: session-id
          in: query
          required: false
          schema:
            type: integer
        - name: session-group-id
          in: query
          required: false
          schema:
            type: integer
        - name: session-id-min
          description: lower bound of an explicit session-id range; requires session-id-max
          in: query
          required: false
          schema:
            type: integer
        - name: session-id-max
          description: upper bound of an explicit session-id range; requires session-id-min
          in: query
          required: false
          schema:
            type: integer
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sessionsResponse'
        404:
          description: not found, instance does not exist
          content:
            text/plain:
              schema:
                type: string
        412:
          description: precondition failed, if instance is not running
          content:
            text/plain:
              schema:
                type: string
        500:
          description: internal server error
          content:
            text/plain:
              schema:
                type: string
  /api/v1/instances/{instance_name}/_logs:
    get:
      summary: Tail the log of a running instance.
      description: >-
        Poll for log lines appended to the instance's run.log since a previous
        call. Pass the "next_offset" from the previous response as the "offset"
        query parameter to only receive newly appended lines. If the instance was
        never started with logging enabled the response is an empty, EOF result
        rather than an error.
      parameters:
        - name: instance_name
          description: instance name of the bngblaster
          in: path
          required: true
          example: sample
          schema:
            type: string
        - name: offset
          description: byte offset into run.log to read from
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          description: maximum number of bytes to read
          in: query
          required: false
          schema:
            type: integer
            default: 65536
            maximum: 1048576
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/logsResponse'
        404:
          description: not found, instance does not exist
          content:
            text/plain:
              schema:
                type: string
        500:
          description: internal server error
          content:
            text/plain:
              schema:
                type: string
...
  /api/v1/instances/{instance_name}/_stop:
    post:
      summary: Stop an instance
      description: >-
        Sends a stop signal to the instance
      parameters:
        - name: instance_name
          description: instance name of the bngblaster
          in: path
          required: true
          example: sample
          schema:
            type: string
      responses:
        202:
          description: accepted
  /api/v1/instances/{instance_name}/_kill:
    post:
      summary: Kill an instance
      description: >-
        Sends a kill signal to the instance
      parameters:
        - name: instance_name
          description: instance name of the bngblaster
          in: path
          required: true
          example: sample
          schema:
            type: string
      responses:
        202:
          description: accepted
  /api/v1/instances/{instance_name}/_command:
    post:
      summary: Send a command to the ctrl socket of the instance.
      description: >-
        The result of the command is a JSON datagram with a specific code field.
        This code field, if set, specifies the http response code.
      parameters:
        - name: instance_name
          description: instance name of the bngblaster
          in: path
          required: true
          example: sample
          schema:
            type: string
      requestBody:
        description: The config file for the bngblaster instance.
        content:
          application/json:
            schema:
              type: object
              properties:
                command:
                  type: string
                arguments:
                  type: object
                  additionalProperties:
                    oneOf:
                      - type: string
                      - type: integer
            example:
              {
                "command": "igmp-join",
                "arguments": {
                  "outer-vlan": 128,
                  "inner-vlan": 7,
                  "group": "232.1.1.3",
                  "source1": "10.0.0.10",
                  "source2": "10.0.0.11",
                  "source3": "10.0.0.12"
                }
              }
      responses:
        200:
          description: ok, if not other specified in the response of the bngblaster
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/commandResponse'
        400:
          description: bad request, body not parsable
          content:
            text/plain:
              schema:
                type: string
        412:
          description: precondition failed, if instance is not running or not existent
          content:
            text/plain:
              schema:
                type: string
        500:
          description: internal server error
          content:
            text/plain:
              schema:
                type: string
  /api/v1/instances/{instance_name}/_files:
    get:
      summary: List downloadable files.
      description: >-
        List the files present in an instance's config folder, used by the web UI's
        download view.
      parameters:
        - name: instance_name
          description: instance name of the parsable
          in: path
          required: true
          example: sample
          schema:
            type: string
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/instanceFile'
        404:
          description: not found, instance does not exist
...
components:
  schemas:
...
    streamsResponse:
      type: object
      properties:
        total:
          description: total number of streams matching the filters
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/streamSummaryStream'
      example:
        {
          "total": 1,
          "offset": 0,
          "limit": 50,
          "items": [
            {
              "flow-id": 1,
              "name": "stream1",
              "type": "ipv4",
              "sub-type": "raw",
              "direction": "upstream",
              "enabled": true,
              "active": true,
              "verified": true,
              "interface": "bblA",
              "tx-packets": 1000,
              "tx-bytes": 64000,
              "rx-packets": 1000,
              "rx-bytes": 64000,
              "rx-loss": 0,
              "tx-pps": 100,
              "rx-pps": 100,
              "session-id": 1,
              "session-traffic": false
            }
          ]
        }
...
    instanceFile:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
      example:
        {
          "name": "run.log",
          "size": 1024
        }
...